+++ rustbus/src/wire/wrapper_types/unixfd.rs
@@ -48,7 +48,7 @@ impl UnixFdInner {
             );
             //  If swapped_fd == fd then we did a sucessful swap and we actually took the value
             if let Ok(taken_fd) = swapped_fd {
-                Some(taken_fd as i32)
+                Some(taken_fd)
             } else {
                 None
             }
